clocksource: stm-global: Provided support for global timestamping at SoC level.
This driver provides access to a common time reference from all cores
of the S32CC SoCs. It allows any core on the SoC to initialize the
time (including the one running Linux) and it also allows any core to
read the current timestamp.

clocksource: stm-global: Fixed a bug where global
                         timestamp worked incorrectly.

Due to the fact that the read/write functions were the big endian ones,
the input from the user and the output were shown in a malformed way.
This commit replaces the big endian functions with raw read/write
functions.

clocksource:: stm-global: Fixed a bug where the global
                          timestamp was not built.

Due to the fact that the Makefile inside drivers/clocksource entry
coresponding to the universal timesource driver was not prefixed with FSL_
it was not compiling. This commit simply adds the required prefix so that
the driver is automatically generated.

Issue: ALB-5464 ALB-5093 ALB-5464 ALB-8823
Signed-off-by: Dorin Ionita <dorin.ionita@nxp.com>

clocksource: stm: Use devm interface where possible

The usage of devm interface and metadata per driver instance will also
allow having multiple instances of the driver without metadata overlaps.

Issue: ALB-10872
Issue: ALB-9261
Signed-off-by: Ghennadi Procopciuc <ghennadi.procopciuc@nxp.com>