gpio: siul2-s32cc: re-enable gpio support
After unifying the two SIUL2 nodes, the siul2 gpio driver needed to be
adapted to handle both hardware modules.
Reading GPIO's value uses the input buffer.
Even though it is deprecated to set the base for the GPIO driver, the
gpio pin ranges would not match the pinctrl ranges if the base is not
correctly configured. This means that calls to retrieve the pinctrl
driver associated to a GPIO pin might fail.

Issue: ALB-8394
Issue: ALB-10252
Issue: ALB-10876
Issue: ALB-10947
Signed-off-by: Andrei Stefanescu <andrei.stefanescu@nxp.com>
Signed-off-by: Andra-Teodora Ilie <andra.ilie@nxp.com>