arch: s32cc: Add trips for each thermal zone
After the latest changes in thermal framework, each thermal zone should
define trips, otherwise devm_thermal_of_zone_register will return an
error.
Taking into account that each temperature reading for a site that is not
monitored with interrupts takes more than 8ms and, in this time, the
TMU interrupts are disabled, a polling-delay of 60ms is used.

Issue: ALB-10862
Signed-off-by: Larisa Grigore <larisa.grigore@nxp.com>