gpio: s32: replace spin_lock_t with raw_spin_lock_t
The GPIO core takes a raw_spinlock_t which, in case of PREEMPT_RT, is
different than a spinlock_t (which can wait). Because of this, we cannot
take a spinlock_t in the GPIO driver callbacks.

Lockdep notices this issue on a non-PREEMPT_RT because of
CONFIG_PROVE_LOCKING.

Issue: ALB-11307

Signed-off-by: Andrei Stefanescu <andrei.stefanescu@nxp.com>