gpio: scmi: don't dispose irq_mappings
Irq mappings are created in the .to_irq callback. They don't need to be
removed in the `scmi_gpio_free`. They are disposed of when the
gpiochip is removed. Moreover, if an interrupt is freed and
re-requested, it will first try to find if an available irq_mapping
exists and use it if it exists or create a new one otherwise.

Issue: ALB-11575

Signed-off-by: Andrei Stefanescu <andrei.stefanescu@nxp.com>